backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pymupdf==1.23.7
pillow==10.1.0
openai==1.3.8
python-dotenv==1.0.0
pytesseract==0.3.10
numpy==1.26.2
pydantic==2.5.2
